fix/refactor: Database Tables & queries update
Add `CREATE TABLE IF NOE EXISTS` queries
Creater fallBacks for App info on the AdminScreen ( Admin subsystem)
Fix problem with infinite loading on the UsersNewsScreen. That was happening because of state change in the wrong place.
Create separate functions to configure admin's subsystem on init.
Refactor getData function to improve code readability.